--- src/main/java/com/ninedocs/userserver/application/emailverficationcode/dto/EmailVerificationCodeResponse.java
@@ -1,9 +1,0 @@
-package com.ninedocs.userserver.application.emailverficationcode.dto;
-
-import lombok.Data;
-
-@Data
-public class EmailVerificationCodeResponse {
-
-  private String verficationCodeExpiredAt;
-}
