+++ src/main/java/com/ninedocs/userserver/application/emailverificationcode/EmailIsDuplicate.java
@@ -1,10 +1,11 @@
-package com.ninedocs.userserver.application.emailverficationcode;
+package com.ninedocs.userserver.application.emailverificationcode;
 
 import com.ninedocs.userserver.persistence.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
-@Service
+@Component
 @RequiredArgsConstructor
 public class EmailIsDuplicate {
