+++ src/main/java/com/ninedocs/userserver/application/emailverificationcode/exception/DuplicateEmailException.java
@@ -1,6 +1,6 @@
-package com.ninedocs.userserver.application.emailverficationcode.exception;
+package com.ninedocs.userserver.application.emailverificationcode.exception;
 
-import com.ninedocs.userserver.application.exception.CustomException;
+import com.ninedocs.userserver.common.exception.CustomException;
 
 public class DuplicateEmailException extends CustomException {
