+++ src/main/java/com/ninedocs/userserver/persistence/User.java
@@ -29,4 +29,7 @@ public class User {
   @Column(nullable = false)
   private String password;
 
+  @Column(nullable = false)
+  private String nickname;
+
 }
