+++ src/main/java/com/ninedocs/userserver/presentation/EmailVerificationCodeController.java
@@ -1,11 +1,10 @@
 package com.ninedocs.userserver.presentation;
 
-import com.ninedocs.userserver.application.emailverficationcode.EmailVerificationCodeService;
-import com.ninedocs.userserver.application.emailverficationcode.dto.EmailDuplicateCodeRequest;
-import com.ninedocs.userserver.application.emailverficationcode.dto.EmailVerificationCodeResponse;
-import com.ninedocs.userserver.application.response.ApiResponse;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
+import com.ninedocs.userserver.application.emailverificationcode.EmailVerificationCodeService;
+import com.ninedocs.userserver.application.emailverificationcode.dto.EmailDuplicateCodeRequest;
+import com.ninedocs.userserver.application.emailverificationcode.dto.EmailVerificationCodeResponse;
+import com.ninedocs.userserver.common.response.ApiResponse;
+import com.ninedocs.userserver.common.staticmethod.CalculateTime;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -13,7 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-public class EmailVerficationController {
+public class EmailVerificationCodeController {
 
   private final EmailVerificationCodeService emailVerificationCodeService;
 
@@ -23,13 +22,9 @@ public class EmailVerficationController {
     // redis 저장 시간 가져오기
     long startTime = emailVerificationCodeService.sendVerificationCode(emailDuplicateCodeRequest);
     //exfiredat 시간 설정 (startTime + 5분)
-    LocalDateTime startDateTime = LocalDateTime.ofEpochSecond(startTime / 1000, 0,
-        java.time.ZoneOffset.UTC);
-    LocalDateTime expiryDateTime = startDateTime.plusMinutes(5);
-    String formattedTime = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss.SSSSSS")
-        .format(expiryDateTime);
+    String formattedTime = CalculateTime.MakeExpiredAt(startTime, 5);
     EmailVerificationCodeResponse emailVerificationCodeResponse = new EmailVerificationCodeResponse();
-    emailVerificationCodeResponse.setVerficationCodeExpiredAt(formattedTime);
+    emailVerificationCodeResponse.setVerificationCodeExpiredAt(formattedTime);
     return ApiResponse.success(emailVerificationCodeResponse);
   }
